Using floats instead of bignums because of overflow issues.

+++ hps_solver/tile_solver.cpp
@@ -10,8 +10,8 @@ void TileSolver::solveTile(Tile* tile, int iterations) {
 
     for (int y_index = 0; y_index < Constants::TILE_HEIGHT; y_index++) {
         for (int x_index = 0; x_index < Constants::TILE_WIDTH; x_index++) {
-            complex c = {stride.real * x_index + origin.real, 
-                         stride.imag * y_index + origin.imag};
+            complex c = {stride.real.toDouble() * x_index + origin.real.toDouble(),
+                         stride.imag.toDouble() * y_index + origin.imag.toDouble()};
             int solution = solvePixel(c, iterations);
             tile->setPoint(x_index, y_index, 255 - (solution * 8));
         }
